added auth logic

+++ backend/controllers/ordersControllers.js
@@ -1,6 +1,7 @@
+const asyncHandler = require('express-async-handler');
 const Order = require('../models/ordersModel');
 
-const registerOrder = async (req, res) => {
+const registerOrder = asyncHandler(async (req, res) => {
   const { product, user, amount } = req.body;
 
   const data = { product, user, amount };
@@ -12,18 +13,20 @@ const registerOrder = async (req, res) => {
     res.status(201).json({
       message: 'Pedido Registrado',
       _id: order._id,
+      user: req.user.id,
     });
   } else {
     res.status(400);
     throw new Error('No se pudo registrar el pedido');
   }
-};
-const getOrders = async (req, res) => {
+});
+
+const getOrders = asyncHandler(async (req, res) => {
   const orders = await Order.find({ user: req.user.id });
   res.status(200).json(orders);
-};
+});
 
-const updateOrder = async (req, res) => {
+const updateOrder = asyncHandler(async (req, res) => {
   const order = await Order.findById(req.params.id);
 
   if (!order) {
@@ -31,12 +34,22 @@ const updateOrder = async (req, res) => {
     throw new Error('Pedido no encontrada');
   }
 
-  const updatedOrder = await Order.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-  });
-  res.status(200).json(updatedOrder);
-};
-const deleteOrder = async (req, res) => {
+  if (order.user.toString() != req.user.id) {
+    res.status(401);
+    throw new Error('Acceso no autorizado');
+  } else {
+    const updatedOrder = await Order.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+        new: true,
+      }
+    );
+    res.status(200).json(updatedOrder);
+  }
+});
+
+const deleteOrder = asyncHandler(async (req, res) => {
   const order = await Order.findById(req.params.id);
 
   if (!order) {
@@ -44,9 +57,14 @@ const deleteOrder = async (req, res) => {
     throw new Error('Pedido no encontrada');
   }
 
-  order.deleteOne();
-  res.status(200).json({ id: order._id });
-};
+  if (order.user.toString() != req.user.id) {
+    res.status(401);
+    throw new Error('Acceso no autorizado');
+  } else {
+    order.deleteOne();
+    res.status(200).json({ id: order._id });
+  }
+});
 
 module.exports = {
   registerOrder,
